Change CodingSize type methods to Static, some were previously configured as Extension methods. Since there were so many, split them off into their own class ("Coding"). The naming of "Encoding.cs" file seemed 'off' when it mostly consisted of the MemoryBuffer class also, so renamed this file to 'MemoryBuffer.cs' Since it was already a change of these two files, I also tidied up the 'using's for the project (just with Visual Studio quick suggestion)
(Also removed SubArray and Memset from .NET Framework version to align
with .NET Core)
