+++ NET Core/LibUA/MemoryBuffer.cs
@@ -1,45 +1,11 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+using System.Buffers;
 
 namespace LibUA
 {
     namespace Core
     {
-        public static class ArrayHelper
-        {
-            public static T[] SubArray<T>(this T[] Data, int Offset, int Length)
-            {
-                var res = new T[Length];
-                Array.Copy(Data, Offset, res, 0, Length);
-
-                return res;
-            }
-
-            public static void Memset<T>(this T[] Data, T Value)
-            {
-                int BlockSize = 32;
-                int Index = 0;
-
-                int Length = Math.Min(BlockSize, Data.Length);
-                while (Index < Length)
-                {
-                    Data[Index++] = Value;
-                }
-
-                Length = Data.Length;
-                while (Index < Length)
-                {
-                    Buffer.BlockCopy(Data, 0, Data, Index, Math.Min(BlockSize, Length - Index));
-                    Index += BlockSize;
-                    BlockSize *= 2;
-                }
-            }
-        }
-
-        public class MemoryBuffer
+        public class MemoryBuffer : IDisposable
         {
             public bool IsReadOnly { get; protected set; }
 
@@ -132,14 +98,46 @@ namespace LibUA
                 }
             }
 
+            private bool isRented;
             public MemoryBuffer(int Size)
             {
                 Position = 0;
                 IsReadOnly = false;
                 IsFixedCapacity = true;
 
-                Buffer = new byte[Size];
+                Buffer = ArrayPool<byte>.Shared.Rent(Size);
                 Capacity = Allocated = Size;
+                isRented = true;
+            }
+
+            private bool disposed;
+
+            public void Dispose()
+            {
+                if (!isRented)
+                    return;
+
+                Dispose(true);
+                GC.SuppressFinalize(this);
+            }
+
+            protected virtual void Dispose(bool disposing)
+            {
+                if (!disposed)
+                {
+                    if (disposing)
+                    {
+                        // Return the array back to the pool
+                        ArrayPool<byte>.Shared.Return(Buffer);
+                    }
+
+                    disposed = true;
+                }
+            }
+
+            ~MemoryBuffer()
+            {
+                Dispose(false);
             }
 
             public bool EnsureAvailable(int Length, bool readOnly)
@@ -248,41 +246,6 @@ namespace LibUA
                 if (!EnsureAvailable(Size, true)) { return false; }
 
                 Array.Copy(Buffer, Position, Dest, 0, Size);
-                Position += Size;
-
-                return true;
-            }
-
-            public bool Prepend(byte[] Add, int Size)
-            {
-                if (Add == null || Size > Add.Length)
-                {
-                    return false;
-                }
-
-                if (Size == 0)
-                {
-                    return true;
-                }
-
-                if (!EnsureAvailable(Size, false))
-                {
-                    return false;
-                }
-
-                var tmp = new byte[Position];
-                if (Position > 0)
-                {
-                    Array.Copy(Buffer, tmp, Position);
-                }
-
-                Array.Copy(Add, Buffer, Size);
-
-                if (Position > 0)
-                {
-                    Array.Copy(tmp, 0, Buffer, Size, Position);
-                }
-
                 Position += Size;
 
                 return true;
@@ -573,87 +536,6 @@ namespace LibUA
                 return true;
             }
 
-            public int CodingSize(byte v) { return 1; }
-            public int CodingSize(SByte v) { return 1; }
-            public int CodingSize(bool v) { return 1; }
-            public int CodingSize(Int16 v) { return 2; }
-            public int CodingSize(UInt16 v) { return 2; }
-            public int CodingSize(Int32 v) { return 4; }
-            public int CodingSize(UInt32 v) { return 4; }
-            public int CodingSize(Int64 v) { return 8; }
-            public int CodingSize(UInt64 v) { return 8; }
-            public int CodingSize(Single v) { return 4; }
-            public int CodingSize(Double v) { return 8; }
-
-            public int CodingSize(string s) { return CodingSize((int)s.Length) + s.Length; }
-
-            public static int VarLenU32Decode(out UInt32 v, byte[] Src)
-            {
-                v = 0;
-
-                for (int i = 0, n = 0; i < 5; i++, n += 7)
-                {
-                    int curLen = 1 + i;
-                    if (Src.Length < curLen)
-                    {
-                        return -1;
-                    }
-
-                    v |= (uint)(((byte)(Src[i] >> 1) & 0x7F) << n);
-                    if ((Src[i] & 1) != 0)
-                    {
-                        continue;
-                    }
-
-                    return curLen;
-                }
-
-                return 0;
-            }
-
-            public static int VarLenU32Encode(UInt32 v, byte[] Dest)
-            {
-                for (int i = 0, n = 0; i < 5; i++, n += 7)
-                {
-                    if (i == 4 && (v >> n) > 0x7F)
-                    {
-                        // Overflow
-                        break;
-                    }
-
-                    Dest[i] = (byte)(((v >> n) & 0x7F) << 1);
-                    if ((v >> n) < 0x80)
-                    {
-                        return i + 1;
-                    }
-
-                    Dest[i] |= 1;
-                }
-
-                // Overflow
-                return 0;
-            }
-
-            public static int VarLenU32Size(UInt32 v)
-            {
-                for (int i = 0, n = 0; i < 5; i++, n += 7)
-                {
-                    if (i == 4 && (v >> n) > 0x7F)
-                    {
-                        // Overflow
-                        break;
-                    }
-
-                    if ((v >> n) < 0x80)
-                    {
-                        return i + 1;
-                    }
-                }
-
-                // Overflow
-                return 0;
-            }
-
             public int VarLenU32Decode(out UInt32 v)
             {
                 int len = Capacity - Position;
@@ -666,7 +548,7 @@ namespace LibUA
                     Array.Copy(Buffer, Position, tmp, 0, bufLen);
                 }
 
-                int adv = VarLenU32Decode(out v, tmp);
+                int adv = Coding.VarLenU32Decode(out v, tmp);
                 if (adv > 0)
                 {
                     Position += adv;
@@ -679,142 +561,13 @@ namespace LibUA
             {
                 var buffer = new byte[8];
 
-                int adv = VarLenU32Encode(v, buffer);
+                int adv = Coding.VarLenU32Encode(v, buffer);
                 if (adv > 0)
                 {
                     Append(buffer, adv);
                 }
 
                 return adv;
-            }
-
-            public static VariantType GetVariantTypeFromType(Type type)
-            {
-                if (type == typeof(bool)) { return VariantType.Boolean; }
-                if (type == typeof(SByte)) { return VariantType.SByte; }
-                if (type == typeof(Byte)) { return VariantType.Byte; }
-                if (type == typeof(Int16)) { return VariantType.Int16; }
-                if (type == typeof(UInt16)) { return VariantType.UInt16; }
-                if (type == typeof(Int32)) { return VariantType.Int32; }
-                if (type == typeof(UInt32)) { return VariantType.UInt32; }
-                if (type == typeof(Int64)) { return VariantType.Int64; }
-                if (type == typeof(UInt64)) { return VariantType.UInt64; }
-                if (type == typeof(float)) { return VariantType.Float; }
-                if (type == typeof(Double)) { return VariantType.Double; }
-                if (type == typeof(string)) { return VariantType.String; }
-                if (type == typeof(byte[])) { return VariantType.ByteString; }
-                if (type == typeof(NodeId)) { return VariantType.NodeId; }
-                if (type == typeof(QualifiedName)) { return VariantType.QualifiedName; }
-                if (type == typeof(LocalizedText)) { return VariantType.LocalizedText; }
-                if (type == typeof(DateTime)) { return VariantType.DateTime; }
-                if (type == typeof(StatusCode)) { return VariantType.StatusCode; }
-                if (type == typeof(ExtensionObject)) { return VariantType.ExtensionObject; }
-
-                // TODO: Other types
-
-                return VariantType.Null;
-            }
-
-
-            public static VariantType GetVariantTypeFromInstance(object obj)
-            {
-                if (obj is bool) { return VariantType.Boolean; }
-                if (obj is SByte) { return VariantType.SByte; }
-                if (obj is Byte) { return VariantType.Byte; }
-                if (obj is Int16) { return VariantType.Int16; }
-                if (obj is UInt16) { return VariantType.UInt16; }
-                if (obj is Int32) { return VariantType.Int32; }
-                if (obj is UInt32) { return VariantType.UInt32; }
-                if (obj is Int64) { return VariantType.Int64; }
-                if (obj is UInt64) { return VariantType.UInt64; }
-                if (obj is float) { return VariantType.Float; }
-                if (obj is Double) { return VariantType.Double; }
-                if (obj is string) { return VariantType.String; }
-                if (obj is byte[]) { return VariantType.ByteString; }
-                if (obj is NodeId) { return VariantType.NodeId; }
-                if (obj is QualifiedName) { return VariantType.QualifiedName; }
-                if (obj is LocalizedText) { return VariantType.LocalizedText; }
-                if (obj is DateTime) { return VariantType.DateTime; }
-                if (obj is StatusCode) { return VariantType.StatusCode; }
-                if (obj is ExtensionObject) { return VariantType.ExtensionObject; }
-
-                // TODO: Other types
-
-                return VariantType.Null;
-            }
-
-            public static Type GetNetType(VariantType type)
-            {
-                if (type == VariantType.Boolean) { return typeof(bool); }
-                if (type == VariantType.SByte) { return typeof(SByte); }
-                if (type == VariantType.Byte) { return typeof(Byte); }
-                if (type == VariantType.Int16) { return typeof(Int16); }
-                if (type == VariantType.UInt16) { return typeof(UInt16); }
-                if (type == VariantType.Int32) { return typeof(Int32); }
-                if (type == VariantType.UInt32) { return typeof(UInt32); }
-                if (type == VariantType.Int64) { return typeof(Int64); }
-                if (type == VariantType.UInt64) { return typeof(UInt64); }
-                if (type == VariantType.Float) { return typeof(float); }
-                if (type == VariantType.Double) { return typeof(Double); }
-                if (type == VariantType.NodeId) { return typeof(NodeId); }
-                if (type == VariantType.QualifiedName) { return typeof(QualifiedName); }
-                if (type == VariantType.LocalizedText) { return typeof(LocalizedText); }
-                if (type == VariantType.String) { return typeof(String); }
-                if (type == VariantType.ExtensionObject) { return typeof(ExtensionObject); }
-
-                // TODO: Other types
-
-                return null;
-            }
-
-            public int VariantCodingSize(object obj)
-            {
-                int rank = 0;
-                VariantType varType;
-                if (obj is Array && !(obj is byte[]))
-                {
-                    var type = obj.GetType();
-                    rank = type.GetArrayRank();
-                    type = type.GetElementType();
-
-                    varType = GetVariantTypeFromType(type);
-                }
-                else
-                {
-                    varType = GetVariantTypeFromInstance(obj);
-                }
-
-                int size = 0;
-
-                byte mask = (byte)varType;
-                ++size;
-
-                if (rank > 1)
-                {
-                    mask |= 0x40;
-                }
-
-                if (rank >= 1)
-                {
-                    mask |= 0x80;
-                    size += CodingSize((int)((Array)obj).Length);
-
-                    foreach (var value in obj as Array)
-                    {
-                        size += VariantCodingSize(value, mask);
-                    }
-
-                    if (rank > 1)
-                    {
-                        size += CodingSize((int)rank) * (1 + rank);
-                    }
-                }
-                else
-                {
-                    size += VariantCodingSize(obj, mask);
-                }
-
-                return size;
             }
 
             public bool VariantEncode(object obj)
@@ -827,11 +580,11 @@ namespace LibUA
                     rank = type.GetArrayRank();
                     type = type.GetElementType();
 
-                    varType = GetVariantTypeFromType(type);
+                    varType = Coding.GetVariantTypeFromType(type);
                 }
                 else
                 {
-                    varType = GetVariantTypeFromInstance(obj);
+                    varType = Coding.GetVariantTypeFromInstance(obj);
                 }
 
                 byte mask = (byte)varType;
@@ -892,8 +645,7 @@ namespace LibUA
                 {
                     if (!Decode(out int arrLen)) { return false; }
                     if (arrLen < 0) { return false; }
-
-                    Type type = GetNetType((VariantType)(mask & 0x3F));
+                    Type type = Coding.GetNetType((VariantType)(mask & 0x3F));
 
                     var arr = Array.CreateInstance(type, arrLen);
                     for (int i = 0; i < arrLen; i++)
@@ -928,45 +680,6 @@ namespace LibUA
                 }
 
                 return true;
-            }
-
-            private int VariantCodingSize(object obj, byte mask)
-            {
-                int size = 0;
-
-                switch (mask & 0x3F)
-                {
-                    case (int)VariantType.Null: break;
-                    case (int)VariantType.Boolean: size += CodingSize((bool)obj); break;
-                    case (int)VariantType.SByte: size += CodingSize((SByte)obj); break;
-                    case (int)VariantType.Byte: size += CodingSize((Byte)obj); break;
-                    case (int)VariantType.Int16: size += CodingSize((Int16)obj); break;
-                    case (int)VariantType.UInt16: size += CodingSize((UInt16)obj); break;
-                    case (int)VariantType.Int32: size += CodingSize((Int32)obj); break;
-                    case (int)VariantType.UInt32: size += CodingSize((UInt32)obj); break;
-                    case (int)VariantType.Int64: size += CodingSize((Int64)obj); break;
-                    case (int)VariantType.UInt64: size += CodingSize((UInt64)obj); break;
-                    case (int)VariantType.Float: size += CodingSize((Single)obj); break;
-                    case (int)VariantType.Double: size += CodingSize((Double)obj); break;
-                    case (int)VariantType.String: size += this.CodingSizeUAString((string)obj); break;
-                    case (int)VariantType.DateTime: size += CodingSize((Int64)0); break;
-                    //case (int)VariantType.Guid: size += CodingSize((int)obj); break;
-                    case (int)VariantType.ByteString: size += this.CodingSizeUAByteString((byte[])obj); break;
-                    //case (int)VariantType.XmlElement: size += CodingSize((int)obj); break;
-                    case (int)VariantType.NodeId: size += this.CodingSize((NodeId)obj); break;
-                    //case (int)VariantType.ExpandedNodeId: size += CodingSize((int)obj); break;
-                    case (int)VariantType.StatusCode: size += CodingSize((UInt32)obj); break;
-                    case (int)VariantType.QualifiedName: size += this.CodingSize((QualifiedName)obj); break;
-                    case (int)VariantType.LocalizedText: size += this.CodingSize((LocalizedText)obj); break;
-                    case (int)VariantType.ExtensionObject: size += this.CodingSize((ExtensionObject)obj); break;
-                    //case (int)VariantType.DataValue: size += CodingSize((int)obj); break;
-                    //case (int)VariantType.Variant: size += CodingSize((int)obj); break;
-                    //case (int)VariantType.DiagnosticInfo: size += CodingSize((int)obj); break;
-                    default:
-                        throw new Exception("TODO");
-                }
-
-                return size;
             }
 
             public bool VariantEncode(object obj, byte mask)
