+++ NET/LibUA/Application.cs
@@ -5,9 +5,7 @@ using System.Linq;
 using System.Net;
 using System.Security.Cryptography;
 using System.Security.Cryptography.X509Certificates;
-using System.Text;
 using System.Threading;
-using System.Threading.Tasks;
 using LibUA.Core;
 
 namespace LibUA
